+++ core/src/main/java/uk/ac/ucl/rits/inform/datasinks/emapstar/repos/LocationVisitAuditRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
  * Audit hospital visit repository.
  * @author Stef Piatek
  */
-public interface LocationVisitAuditRepository extends CrudRepository<LocationVisitAudit, Integer> {
+public interface LocationVisitAuditRepository extends CrudRepository<LocationVisitAudit, Long> {
     /**
      * For testing: find by location string.
      * @param location full location string
